requirements.txt
azure-functions
pyqrcode
pypng
